+++ Taller_TIC1_Backend/Models/Factura.cs
@@ -5,9 +5,9 @@ namespace Taller_TIC1_Backend.Models
         public int Id { get; set; }
         public int IdOrdenTrabajo { get; set; }
         public DateTime FechaEmision { get; set; }
-        public float Subtotal { get; set; }
-        public float Impuestos { get; set; }
-        public float Total { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal Impuestos { get; set; }
+        public decimal Total { get; set; }
         public string Estado { get; set; } = string.Empty; // "Pendiente", "Pagada", "Cancelada"
         
         // Navegación
