+++ Taller_TIC1_Backend/Models/Repuesto.cs
@@ -5,7 +5,7 @@ namespace Taller_TIC1_Backend.Models
         public int Id { get; set; }
         public string Nombre { get; set; } = string.Empty;
         public long Numero_serie { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
         public int Stock { get; set; }
     }
 }
