+++ Taller_TIC1_Backend/Models/Servicio.cs
@@ -9,7 +9,7 @@ namespace Taller_TIC1_Backend.Models
         public int IdVehiculo { get; set; }
         public int? IdEmpleado { get; set; }
         public int IdEstado { get; set; }
-        public float Costo { get; set; }
+        public decimal Costo { get; set; }
         public DateTime FechaCreacion { get; set; } = DateTime.UtcNow;
         public DateTime? FechaActualizacion { get; set; }
